Refactor: mystudy 화면 코드 개선 및 패키지 구조 변경
