+++ app/src/main/java/com/sesac/developer_study_platform/ui/category/CategoryFragment.kt
@@ -12,8 +12,8 @@ import com.sesac.developer_study_platform.Category
 import com.sesac.developer_study_platform.R
 import com.sesac.developer_study_platform.data.source.remote.StudyService
 import com.sesac.developer_study_platform.databinding.FragmentCategoryBinding
-import com.sesac.developer_study_platform.ui.GridSpaceItemDecoration
-import com.sesac.developer_study_platform.ui.StudyClickListener
+import com.sesac.developer_study_platform.ui.common.GridSpaceItemDecoration
+import com.sesac.developer_study_platform.ui.common.StudyClickListener
 import kotlinx.coroutines.launch
 
 class CategoryFragment : Fragment() {
