+++ app/src/main/java/com/sesac/developer_study_platform/ui/common/StudyAdapter.kt
@@ -1,4 +1,4 @@
-package com.sesac.developer_study_platform.ui.home
+package com.sesac.developer_study_platform.ui.common
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,7 +7,6 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sesac.developer_study_platform.data.UserStudy
 import com.sesac.developer_study_platform.databinding.ItemStudyBinding
-import com.sesac.developer_study_platform.ui.StudyClickListener
 import com.sesac.developer_study_platform.util.getAllDayList
 import com.sesac.developer_study_platform.util.getDayList
 import com.sesac.developer_study_platform.util.setImage
