+++ app/src/main/java/com/sesac/developer_study_platform/ui/common/StudyClickListener.kt
@@ -1,4 +1,4 @@
-package com.sesac.developer_study_platform.ui
+package com.sesac.developer_study_platform.ui.common
 
 interface StudyClickListener {
