+++ app/src/main/java/com/sesac/developer_study_platform/ui/searchresult/SearchAdapter.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.sesac.developer_study_platform.R
 import com.sesac.developer_study_platform.data.Study
 import com.sesac.developer_study_platform.databinding.ItemSearchBinding
-import com.sesac.developer_study_platform.ui.StudyClickListener
+import com.sesac.developer_study_platform.ui.common.StudyClickListener
 import com.sesac.developer_study_platform.util.setImage
 
 class SearchAdapter(private val clickListener: StudyClickListener) :
