+++ app/src/main/java/com/sesac/developer_study_platform/ui/studyform/DayTimeAdapter.kt
@@ -7,7 +7,6 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sesac.developer_study_platform.data.DayTime
 import com.sesac.developer_study_platform.databinding.ItemDayTimeBinding
-import com.sesac.developer_study_platform.ui.DayTimeClickListener
 
 class DayTimeAdapter(private val dayTimeClickListener: DayTimeClickListener) :
     ListAdapter<DayTime, DayTimeAdapter.DayTimeViewHolder>(diffUtil) {
