+++ app/src/main/res/layout/fragment_my_study.xml
@@ -16,9 +16,10 @@
         app:titleTextAppearance="@style/ToolbarText" />
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_study"
+        android:id="@+id/rv_study_list"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        android:layout_marginHorizontal="@dimen/space_median"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
